fix user merge error
The error occurred because the functions "get_all_related_objects" and "get_all_field_names" are deprecated and not returning the exact same items as they used to. Resolved by replacing those two functions with get_fields(). In addition, the function was not handling many-to-many fields. Fixed that as well.

+++ tendenci/apps/profiles/views.py
@@ -1154,14 +1154,16 @@ def merge_profiles(request, sid, template_name="profiles/merge_profiles.html"):
         if form.is_valid():
             master = form.cleaned_data["master_record"]
             users = form.cleaned_data['user_list']
+            
+            master_user = master.user
 
             if master and users:
                 # get description for event log before users get deleted
                 description = 'Master user: %s, merged user(s): %s.' % (
-                                '%s %s (%s)(id=%d)' % (master.user.first_name,
-                                               master.user.last_name,
-                                               master.user.username,
-                                               master.user.id),
+                                '%s %s (%s)(id=%d)' % (master_user.first_name,
+                                               master_user.last_name,
+                                               master_user.username,
+                                               master_user.id),
                                 ', '.join(['%s %s (%s)(id=%d)' % (
                                 profile.user.first_name,
                                 profile.user.last_name,
@@ -1169,56 +1171,74 @@ def merge_profiles(request, sid, template_name="profiles/merge_profiles.html"):
                                 profile.user.id
                                 ) for profile in users if profile != master]))
         
-                related = master.user._meta.get_all_related_objects()
-                field_names = master._meta.get_all_field_names()
+                related = master_user._meta.get_fields()
+                field_names = [field.name for field in master._meta.get_fields()]
         
                 valnames = dict()
                 for r in related:
-                    if not r.model is Profile:
-                        valnames.setdefault(r.model, []).append(r.field)
+                    if not r.related_model is Profile:
+                        if not r.related_model:
+                            continue
+                        valnames.setdefault(r.related_model, []).append(r)
         
                 for profile in users:
+                    user_to_delete = profile.user
                     if profile != master:
                         for field in field_names:
                             if getattr(master, field) == '':
                                 setattr(master, field, getattr(profile, field))
         
                         for model, fields in valnames.iteritems():
-        
+                            # skip auth_user
+                            if model is User:
+                                continue
+
                             for field in fields:
-                                if not isinstance(field, models.OneToOneField):
-                                    objs = model.objects.filter(**{field.name: profile.user})
+                                if isinstance(field, models.ManyToManyField):
+                                    # handle ManyToMany
+                                    items = eval('user_to_delete.%s.all()' % field.name)
+                                    if items:
+                                        for item in items:
+                                            # add to master_user
+                                            eval('master_user.%s.add(item)' % field.name)
+                                        # clear from user_to_delete
+                                        eval('user_to_delete.%s.clear()' % field.name)
+                                    continue
+                                
+                                field_name = field.field.name
+                                if not isinstance(field, models.OneToOneField) and not isinstance(field, models.OneToOneRel):
+                                    objs = model.objects.filter(**{field_name: user_to_delete})
         
                                     # handle unique_together fields. for example, GroupMembership
                                     # unique_together = ('group', 'member',)
                                     [unique_together] = model._meta.unique_together[:1] or [None]
-                                    if unique_together and field.name in unique_together:
+                                    if unique_together and field_name in unique_together:
                                         for obj in objs:
-                                            field_values = [getattr(obj, field_name) for field_name in unique_together]
+                                            field_values = [getattr(obj, fieldname) for fieldname in unique_together]
                                             field_dict = dict(zip(unique_together, field_values))
                                             # switch to master user
-                                            field_dict[field.name] = master.user
+                                            field_dict[field_name] = master_user
                                             # check if the master record exists
                                             if model.objects.filter(**field_dict).exists():
                                                 obj.delete()
                                             else:
-                                                setattr(obj, field.name, master.user)
+                                                setattr(obj, field_name, master_user)
                                                 obj.save()
                                     else:
                                         if objs.exists():
                                             try:
-                                                objs.update(**{field.name: master.user})
+                                                objs.update(**{field_name: master_user})
                                             except Exception:
                                                 connection._rollback()
                                 else:  # OneToOne
-                                    [obj] = model.objects.filter(**{field.name: profile.user})[:1] or [None]
+                                    [obj] = model.objects.filter(**{field_name: user_to_delete})[:1] or [None]
                                     if obj:
-                                        [master_obj] = model.objects.filter(**{field.name: master.user})[:1] or [None]
+                                        [master_obj] = model.objects.filter(**{field_name: master_user})[:1] or [None]
                                         if not master_obj:
-                                            setattr(obj, field.name, master.user)
+                                            setattr(obj, field_name, master_user)
                                             obj.save()
                                         else:
-                                            obj_fields = master_obj._meta.get_all_field_names()
+                                            obj_fields = [field.name for field in master_obj._meta.get_fields()]
                                             updated = False
                                             for fld in obj_fields:
                                                 master_val = getattr(master_obj, fld)
@@ -1233,8 +1253,9 @@ def merge_profiles(request, sid, template_name="profiles/merge_profiles.html"):
                                             obj.delete()
         
                         master.save()
-                        profile.user.delete()
                         profile.delete()
+                        user_to_delete.delete()
+                        
         
                 # log an event
                 EventLog.objects.log(description=description[:120])
